# backend/openapi/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    ClusterReadiness:
      type: object
      properties:
        clusterConnected:
          type: boolean
        targetClusterId:
          type: string
        connectionStatus:
          type: string
        prerequisites:
          type: array
          items:
            $ref: "#/components/schemas/MigrationPrerequisite"
...
    MigrationPrerequisite:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
        title:
          type: string
        description:
          type: string
        requiredByPlan:
          type: boolean
        optionalTier:
          type: boolean
        docUrl:
          type: string
        status:
          type: string
        triggeredByCount:
          type: integer
          format: int32
...
    ProjectInfo:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        creationTimestamp:
          type: string
        hasThreeScale:
          type: boolean
        hasKuadrant:
          type: boolean
...
    TargetCluster:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        apiServerUrl:
          type: string
        token:
          type: string
        authType:
          type: string
        verifySsl:
          type: boolean
        enabled:
          type: boolean
    TargetClusterView:
      description: Target Kubernetes cluster (public view — credentials never exposed)
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        apiServerUrl:
          type: string
        authType:
          type: string
        verifySsl:
          type: boolean
        enabled:
          type: boolean
        credentialConfigured:
          type: boolean
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      description: Authentication
paths:
...
  /api/cluster/projects/{name}:
    get:
      operationId: getProject
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectInfo"
      summary: Get Project
      tags:
      - Cluster Resource
  /api/cluster/readiness:
    get:
      operationId: getReadiness
      parameters:
      - name: planId
        in: query
        schema:
          type: string
      - name: targetClusterId
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterReadiness"
      summary: Get Readiness
      tags:
      - Cluster Resource
  /api/cluster/targets:
    get:
      operationId: listTargetClusters
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TargetClusterView"
      summary: List Target Clusters
      tags:
      - Cluster Resource
    post:
      operationId: addTargetCluster
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetCluster"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TargetClusterView"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Add Target Cluster
      tags:
      - Cluster Resource
      security:
      - SecurityScheme:
        - admin
  /api/cluster/targets/{id}:
    delete:
      operationId: removeTargetCluster
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Remove Target Cluster
      tags:
      - Cluster Resource
      security:
      - SecurityScheme:
        - admin
  /api/cluster/targets/{id}/validate:
    get:
      operationId: validateTargetCluster
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
      summary: Validate Target Cluster
      tags:
      - Cluster Resource
...
info:
  title: ApiShift API
  version: 0.3.0
  description: 3scale to Connectivity Link migration console REST API
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value